requirements.txt
aiogram==3.24.0
aiosqlite==0.22.1
Pillow==10.0.0
